# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediNova API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReminderFrequency =
  (typeof ReminderFrequency)[keyof typeof ReminderFrequency];

export const ReminderFrequency = {
  daily: "daily",
  twice_daily: "twice_daily",
  three_times_daily: "three_times_daily",
  weekly: "weekly",
  as_needed: "as_needed",
} as const;

export interface Reminder {
  id: number;
  medicationName: string;
  dosage: string;
  frequency: ReminderFrequency;
  /** Array of HH:MM strings */
  times: string[];
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  isActive: boolean;
  color?: string | null;
  createdAt: string;
}

export type CreateReminderBodyFrequency =
  (typeof CreateReminderBodyFrequency)[keyof typeof CreateReminderBodyFrequency];

export const CreateReminderBodyFrequency = {
  daily: "daily",
  twice_daily: "twice_daily",
  three_times_daily: "three_times_daily",
  weekly: "weekly",
  as_needed: "as_needed",
} as const;

export interface CreateReminderBody {
  medicationName: string;
  dosage: string;
  frequency: CreateReminderBodyFrequency;
  times: string[];
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  color?: string | null;
}

export type UpdateReminderBodyFrequency =
  (typeof UpdateReminderBodyFrequency)[keyof typeof UpdateReminderBodyFrequency];

export const UpdateReminderBodyFrequency = {
  daily: "daily",
  twice_daily: "twice_daily",
  three_times_daily: "three_times_daily",
  weekly: "weekly",
  as_needed: "as_needed",
} as const;

export interface UpdateReminderBody {
  medicationName?: string;
  dosage?: string;
  frequency?: UpdateReminderBodyFrequency;
  times?: string[];
  startDate?: string;
  endDate?: string | null;
  notes?: string | null;
  isActive?: boolean;
  color?: string | null;
}

export type LogDoseBodyStatus =
  (typeof LogDoseBodyStatus)[keyof typeof LogDoseBodyStatus];

export const LogDoseBodyStatus = {
  taken: "taken",
  missed: "missed",
  skipped: "skipped",
} as const;

export interface LogDoseBody {
  status: LogDoseBodyStatus;
  takenAt?: string | null;
  scheduledTime: string;
}

export type DoseLogStatus = (typeof DoseLogStatus)[keyof typeof DoseLogStatus];

export const DoseLogStatus = {
  taken: "taken",
  missed: "missed",
  skipped: "skipped",
} as const;

export interface DoseLog {
  id: number;
  reminderId: number;
  status: DoseLogStatus;
  scheduledTime: string;
  takenAt?: string | null;
  createdAt: string;
}

export type TodayReminderStatus =
  (typeof TodayReminderStatus)[keyof typeof TodayReminderStatus];

export const TodayReminderStatus = {
  pending: "pending",
  taken: "taken",
  missed: "missed",
  skipped: "skipped",
} as const;

export interface TodayReminder {
  id: number;
  medicationName: string;
  dosage: string;
  scheduledTime: string;
  status: TodayReminderStatus;
  color?: string | null;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  language?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
  /** ISO language code (en, hi, te, ta, ml, etc.) */
  language?: string | null;
}

export interface ChatResponse {
  userMessage: ChatMessage;
  assistantMessage: ChatMessage;
  detectedLanguage?: string | null;
}

export interface DashboardSummary {
  totalReminders: number;
  activeReminders: number;
  todayTotal: number;
  todayTaken: number;
  todayMissed: number;
  /** Percentage 0-100 */
  adherenceRate: number;
  /** Consecutive days of full adherence */
  streak: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  dose_taken: "dose_taken",
  dose_missed: "dose_missed",
  reminder_created: "reminder_created",
  reminder_updated: "reminder_updated",
  chat_message: "chat_message",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  medicationName?: string | null;
  createdAt: string;
}

export interface AdherenceDay {
  date: string;
  taken: number;
  missed: number;
  total: number;
  rate: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type GetChatHistoryParams = {
  limit?: number;
};

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};